refactor(vms/saevm/cchain): tidy MinPriceExponent wiring per review
